# docs/swagger.yaml
definitions:
  handler.createUserRequest:
    properties:
      username:
        type: string
    type: object
  handler.userResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: gohex – simple Go framework to build APIs
  title: gohex API
  version: "1.0"
paths:
  /users:
    post:
      consumes:
      - core/json
      description: Create a user with a username
      parameters:
      - description: Create user payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequest'
      produces:
      - core/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - core/json
      description: Retrieve a user by its ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - core/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Provide your Bearer token in the format: ''Bearer {token}'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"